feat: Update dependencies and use MaterialExpressiveTheme
This commit updates several dependencies to their newer versions:
- Gradle to 8.14.2
- Android Gradle Plugin to 8.11.0
- androidxComposeBom to 2025.06.01
- androidxCore to 1.16.0
- androidxActivity to 1.10.1
- androidxLifecycle to 2.9.1
- material3 to 1.4.0-alpha16
- Kotlin to 2.2.0

It also migrates the app theme from `MaterialTheme` to `MaterialExpressiveTheme` and updates the `compileSdk` and `targetSdk` to 36.
The app version code has been incremented to 5, and the version patch to 4.
Additionally, copying text to the clipboard now uses `LocalClipboard.nativeClipboard` and `ClipData`.